TERMINADA_Biblioteca_Con_Python/library_api/requirements.txt
Django>=4.0,<5.0
djangorestframework>=3.14,<4.0
djangorestframework-simplejwt>=5.0,<6.0
